# backend/docs/apifox_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Canvas OS API",
    "description": "\n# Life Canvas OS 后端 API\n\n八维生命平衡系统后端接口，支持个人成长管理、日记记录、AI 洞察等功能。\n\n## 主要功能\n\n- **用户认证**: PIN 码设置、验证、修改\n- **系统管理**: 八维系统（饮食、运动、智力、输出、恢复、资产、连接、环境）\n- **用户配置**: 用户信息、设置、AI 配置\n- **日记管理**: 日记 CRUD、附件、编辑历史\n- **AI 洞察**: 基于用户数据生成个性化洞察\n- **数据管理**: 备份、恢复、导出、导入\n\n## 认证方式\n\n本 API 使用 PIN 码认证。\n\n## 响应格式\n\n所有接口遵循统一的响应格式：\n\n```json\n{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1707219200000\n}\n```\n\n## 错误码\n\n- `200`: 成功\n- `201`: 创建成功\n- `400`: 错误请求\n- `401`: 未授权\n- `404`: 资源不存在\n- `409`: 冲突\n- `422`: 参数验证失败\n- `429`: 请求过于频繁\n- `500`: 服务器错误\n\n## 分页参数\n\n| 参数 | 类型 | 必填 | 默认值 | 说明 |\n|------|------|------|--------|------|\n| page | Integer | 否 | 1 | 页码，从 1 开始 |\n| page_size | Integer | 否 | 20 | 每页数量，最大 100 |\n| sort_by | String | 否 | created_at | 排序字段 |\n| sort_order | String | 否 | desc | 排序方向：asc/desc |\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/systems": {
      "get": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "获取系统列表",
        "description": "获取所有系统列表",
        "operationId": "get_systems_api_systems_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "score",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SystemResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        },
                        "has_next": {
                          "type": "boolean"
                        },
                        "has_prev": {
                          "type": "boolean"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/{system_type}": {
      "get": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "获取系统详情",
        "description": "获取系统详情",
        "operationId": "get_system_detail_api_systems__system_type__get",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/{system_type}/score": {
      "patch": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "更新系统评分",
        "description": "更新系统评分",
        "operationId": "update_system_score_api_systems__system_type__score_patch",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemScoreUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemScoreUpdateResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/{system_type}/logs": {
      "post": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "添加日志",
        "description": "添加日志",
        "operationId": "create_system_log_api_systems__system_type__logs_post",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemLogCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemLogResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "获取日志列表",
        "description": "获取日志列表",
        "operationId": "get_system_logs_api_systems__system_type__logs_get",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SystemLogResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/{system_type}/actions": {
      "post": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "添加行动项",
        "description": "添加行动项",
        "operationId": "create_system_action_api_systems__system_type__actions_post",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemActionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemActionResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/{system_type}/actions/{action_id}": {
      "patch": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "更新行动项",
        "description": "更新行动项",
        "operationId": "update_system_action_api_systems__system_type__actions__action_id__patch",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          },
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemActionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemActionResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "systems",
          "systems"
        ],
        "summary": "删除行动项",
        "description": "删除行动项",
        "operationId": "delete_system_action_api_systems__system_type__actions__action_id__delete",
        "parameters": [
          {
            "name": "system_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Type"
            }
          },
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemActionDeleteResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SystemActionCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "title": "Text",
            "description": "行动项内容"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SystemActionCreate",
        "description": "创建系统行动项"
      },
      "SystemActionUpdate": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed"
          }
        },
        "type": "object",
        "title": "SystemActionUpdate",
        "description": "更新系统行动项"
      },
      "SystemLogCreate": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 100,
            "title": "Label",
            "description": "日志标签"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "日志内容"
          },
          "meta_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Data",
            "description": "额外元数据"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "SystemLogCreate",
        "description": "创建系统日志"
      },
      "SystemScoreUpdate": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "SystemScoreUpdate",
        "description": "更新系统评分"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "SystemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "FUEL",
              "PHYSICAL",
              "INTELLECTUAL",
              "OUTPUT",
              "RECOVERY",
              "ASSET",
              "CONNECTION",
              "ENVIRONMENT"
            ]
          },
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "details": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SystemScoreUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "old_score": {
            "type": "integer"
          },
          "new_score": {
            "type": "integer"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SystemLogResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "system_id": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "meta_data": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SystemActionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "system_id": {
            "type": "integer"
          },
          "text": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SystemActionDeleteResponse": {
        "type": "object",
        "properties": {
          "deleted_id": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境"
    }
  ],
...
}